+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change route ordering. `/health` route is hard-configured as first route to be considered ([PORTAL-859](https://issue.inventage.com/browse/PORTAL-859)).
 - Add stacktrace to logback configuration ([PORTAL-1243](https://issue.inventage.com/browse/PORTAL-1243)).
 - [PORTAL-1290](https://issue.inventage.com/browse/PORTAL-1290) : support mounting of gateway-routing-config files via configMap. The directory `proxy-config.examples` contains a selected set of example config files for each microservice. Consumers of this chart should selectively copy these files and adapt according to their specific environment.
-- New optional `pathsWithoutSessionTimeoutReset` parameter in session middleware added. There will be no session timeout reset if a request uri matches configured regex ([PORTAL-1321](https://issue.inventage.com/browse/PORTAL-1321)).
+- New optional `uriWithoutSessionTimeoutReset` parameter in session middleware added. There will be no session timeout reset if a request contains one of the configured paths ([PORTAL-1321](https://issue.inventage.com/browse/PORTAL-1321))
 
 ## [7.0.2]-202304141241-4-576b35e4 - 2023-04-17